simulation/requirements.txt
psycopg2-binary>=2.9
prometheus_client>=0.17